+++ app/src/components/MachineNumber/index.js
@@ -1,129 +1,7 @@
 import React from 'react';
 import _ from 'lodash';
 import gnum from'../../helpers/global-enum';
-
-class PartNumber extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.quantity = 0;
-    this.updateForm = this.updateForm.bind(this);
-    this.handleSelectChange = this.handleSelectChange.bind(this);
-    this.renderSelect = this.renderSelect.bind(this);
-    this.subtractOne = this.subtractOne.bind(this);
-    this.addOne = this.addOne.bind(this);
-    this.renderQuantity = this.renderQuantity.bind(this);
-    this.state = {
-      parentValue: this.props.parentValue,
-      choiceValue: _.clone(this.props.value),
-      quantity: 1,
-      timestamp: Date.now()
-    };
-  }
-
-  componentWillUpdate(nextProps, nextState) {
-    console.log('Part Number componentWillUpdate');
-    console.log(nextState.quantity);
-    console.log(nextState.choiceValue);
-    console.log('nextProps.parentValue');
-    if (this.props.parentValue !== nextProps.parentValue) {
-      this.setState({
-        ...this.state,
-        choiceValue: 'default'
-      });
-    }
-    console.log(nextProps.parentValue);
-    console.log('this.props.parentValue');
-    console.log(this.props.parentValue);
-
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    // re-render the ItemForm component with new form data
-    // check if timestamp hasn't been updated in the past
-    // .01 seconds so that infinite loop doesn't occur
-    if (Date.now() - this.state.timestamp > 10) {
-      console.log('Part Number Component updating');
-      return true;
-    }
-    return false;
-  }
-
-  handleSelectChange(e) {
-    const partNumber = e.target.options[e.target.selectedIndex].value;
-      this.setState({
-        ...this.state,
-        choiceValue: partNumber,
-        quantity: partNumber !== 'default' ? 1 : undefined,
-        timestamp: Date.now()
-      }, this.updateForm);
-  }
-
-  renderSelect(props) {
-    const partNumbers = this.props.partNumbers;
-    return (
-      <select
-        id={this.props.id}
-        value={this.state.choiceValue}
-        onChange={this.handleSelectChange}>
-        <option value="default">--part number--</option>
-        {
-          partNumbers.map(partNum => (
-            <option key={partNum} value={partNum}>{partNum}</option>
-          ))
-        }
-      </select>
-    )
-  }
-
-  updateForm() {
-    if (this.state.choiceValue !== 'default') {
-      this.props.updateForm(this.props.index, { partNumber: this.state.choiceValue, quantity: this.state.quantity
-      });
-    }
-  }
-
-  subtractOne(e) {
-    e.preventDefault();
-    const currentQuantity = this.state.quantity;
-    // to remove a part entirely, we'll use a remove button
-    if (currentQuantity > 1) {
-      this.setState({
-        ...this.state,
-        quantity: currentQuantity - 1
-      }, this.updateForm);
-    }
-  }
-
-  addOne(e) {
-    e.preventDefault();
-    this.setState({
-      ...this.state,
-      quantity: this.state.quantity + 1
-    }, this.updateForm);
-  }
-
-  renderQuantity() {
-    return (
-      <div className="quantity">
-        <span>{this.state.quantity}</span>
-        <button onClick={this.subtractOne}>-</button>
-        <button onClick={this.addOne}>+</button>
-      </div>
-    )
-  }
-
-  render() {
-    console.log('this.state.choiceValue');
-    console.log(this.state.choiceValue);
-    return (
-      <div key={this.state.timestamp} className="PartNumber" style={{border: 'solid 1px red'}}>
-        {this.renderSelect()}
-        {this.state.choiceValue === 'default' ? null : this.renderQuantity()}
-      </div>
-    )
-  }
-}
+import PartNumber from '../PartNumber';
 
 class MachineNumber extends React.Component {
   constructor(props) {
